refactor(api): migrate CopilotDataFacade to common module
