+++ src/qdash/api/dependencies.py
@@ -19,7 +19,6 @@ from qdash.api.services.config_service import ConfigService
 from qdash.api.services.cooldown_service import CooldownService
 from qdash.api.services.cooldown_wiring_event_service import CooldownWiringEventService
 from qdash.api.services.copilot_chat_session_service import CopilotChatSessionService
-from qdash.api.services.copilot_data_facade import CopilotDataFacade
 from qdash.api.services.cryostat_service import CryostatService
 from qdash.api.services.device_topology_service import DeviceTopologyService
 from qdash.api.services.execution_service import ExecutionService
@@ -38,6 +37,7 @@ from qdash.api.services.seed_import_service import SeedImportService
 from qdash.api.services.task_file_service import TaskFileService
 from qdash.api.services.task_result_service import TaskResultService
 from qdash.api.services.task_service import TaskService
+from qdash.common.copilot.data_facade import CopilotDataFacade
 from qdash.repository import (
     MongoChipRepository,
     MongoExecutionCounterRepository,
