+++ src/qdash/api/routers/copilot.py
@@ -35,12 +35,12 @@ from qdash.api.schemas.copilot_chat_session import (
 from qdash.api.services.copilot_chat_session_service import (
     CopilotChatSessionService,
 )
-from qdash.api.services.copilot_data_facade import CopilotDataFacade
 from qdash.common.copilot.analysis_models import (
     AnalysisResponse,
     AnalyzeRequest,
     ChatRequest,
 )
+from qdash.common.copilot.data_facade import CopilotDataFacade
 from qdash.common.copilot.settings import CopilotConfig, ModelConfig, load_copilot_config
 from qdash.datamodel.task_knowledge import get_task_knowledge
