+++ src/qdash/api/routers/forum.py
@@ -280,7 +280,7 @@ async def forum_ai_reply_stream(
             body.user_message,
         )
 
-        from qdash.api.services.copilot_data_facade import CopilotDataFacade
+        from qdash.common.copilot.data_facade import CopilotDataFacade
 
         copilot_data_svc = CopilotDataFacade()
         tool_executors = copilot_data_svc.build_tool_executors()
