+++ src/qdash/api/routers/issue.py
@@ -328,7 +328,7 @@ async def issue_ai_reply_stream(
         await asyncio.sleep(0)
 
         # Build tool executors
-        from qdash.api.services.copilot_data_facade import CopilotDataFacade
+        from qdash.common.copilot.data_facade import CopilotDataFacade
 
         copilot_data_svc = CopilotDataFacade()
         tool_executors = copilot_data_svc.build_tool_executors()
