+++ src/qdash/api/services/task_result_service.py
@@ -671,7 +671,7 @@ class TaskResultService:
         config: Any,
     ) -> Any:
         """Build the compact context passed to AI triage analysis."""
-        from qdash.api.services.copilot_data_facade import CopilotDataFacade
+        from qdash.common.copilot.data_facade import CopilotDataFacade
 
         context_bundle = CopilotDataFacade().build_analysis_context(
             task_name=task_name,
